Jetpack: Sharing: Added filter to allow sharing buttons to show on multiple the_content filters
By default the sharing buttons where not displaying if the_content was in global $wp_current_filter
more than one time. This was added as an edge case to catch a specific theme attempting to load
a featured image from the content that belonged to Pintrest. That behavior was also causing 3rd
party plugins that rely on filtering the_content to lose the sharing buttons. There is now a
filter, jetpack_sharing_infinite_catch, that can be enabled if this infinite issue shows again.