First attempt at swapping to a concatenated js pkg
This is dynamically switching them on/off.

We’ll probably need to reorder them if some dependencies internally got
out of sequence in the concatenated JS.

Also need to hit some items that are getting printed out after the
concatenated one is slated to hit — the flags for them being included
aren’t getting flagged.

cc: @rase- — this is what I meant to show you on Tuesday, but kept
forgetting.