Initial commit of Security Center class.
This class adds an admin page under Jetpack with an alert if any of the
three areas aren’t covered.

It needs some serious style love, but it’s there, and there are some
classes wrapping the content already.

I may have overcomplicated things a bit with an extra set of actions
triggering things from the admin page, rather than putting what it
calls there directly — may simplify in a future commit, but it also
makes it easier to play with alternate interfaces while in development.